refactor(store): decompose appReducer god-switch into per-domain slices
TS1 from the 2026-05-10 codebase review (issue #275). The 838-LOC
`appReducer` switch with ~50 cases (and a 1127-LOC test file as the
tell) is gone. Now `apps/web/src/renderer/lib/store/reducer.ts` is a
1-line re-export and the implementation lives in
`apps/web/src/renderer/lib/store/reducers/`:

- `messagesReducer.ts` — chat-message and compose actions; also owns
  the exported `handleChatEvent`
- `conversationReducer.ts` — conversation history + hydration
- `mindsReducer.ts` — minds + models + active-mind selection
- `lifecycleReducer.ts` — landing, account switching, runtime phase
- `a2aReducer.ts` — A2A messages + task updates
- `chatroomReducer.ts` — every CHATROOM_* and orchestration config
- `helpers.ts` — shared utilities (selectedModelForActiveMind,
  setConversationView, mergeConversationSummaries, getPlainContent, etc.)
- `index.ts` — dispatcher: Record<AppAction['type'], (state, action)
  => Partial<AppState> | AppState> with reference-identity short-circuit
  for no-op handlers

The dispatcher merges Partial<AppState> returns via { ...state, ...next }
and skips the spread when a handler returns the same state reference
(no-op semantics preserved).

Test reorganization: `reducer.test.ts` (1127 LOC, 111 cases) kept as
the integration safety net rather than split per-slice. Splitting the
tests is a separate cleanup; the integration coverage is what protects
against the kind of behavior drift the move surfaced (initial extraction
got 5 handleChatEvent cases wrong — tool_done result append, reasoning
append-to-existing-block, message_final no-op-on-same-sdkMessageId,
error/timeout text format, the timeoutMs/1000 divisor — all caught and
fixed by the test suite before this commit).

LOC budget: reducer.ts is 1 line (was 838); per-slice files range
59–274 LOC. messagesReducer.ts and chatroomReducer.ts exceed the issue's
<150 LOC suggestion because of single large action handlers
(handleChatEvent ~100 LOC, CHATROOM_EVENT orchestration switch ~100
LOC). Further splitting those into their own files is a Boy Scout
follow-up that would not change behavior.

Bumps to v0.52.3 (PR #272/#273 bump to v0.52.1, PR #283 to v0.52.2).

Closes #275
