+++ apps/web/src/renderer/lib/store/reducer.ts
@@ -1,864 +1,1 @@
-import type { ChatMessage, ChatEvent, ContentBlock, ConversationSummary } from '@chamber/shared/types';
-import type { Task, TaskState } from '@chamber/shared/a2a-types';
-import type { ChatroomMessage, TaskLedgerItem } from '@chamber/shared/chatroom-types';
-import { isOrchestrationEvent } from '@chamber/shared/chatroom-types';
-import type { AppState, AppAction, ConversationViewState } from './state';
-
-/** Extract plain text from content blocks (for search, accessibility, etc.) */
-export function getPlainContent(message: ChatMessage): string {
-  return message.blocks
-    .filter((b): b is Extract<ContentBlock, { type: 'text' }> => b.type === 'text')
-    .map(b => b.content)
-    .join('');
-}
-
-function nonEmptyString(value: unknown, fallback: string): string {
-  return typeof value === 'string' && value.trim().length > 0 ? value.trim() : fallback;
-}
-
-function updateChatMessage<T extends ChatMessage>(
-  message: T,
-  updates: Partial<Pick<ChatMessage, 'blocks' | 'isStreaming'>>,
-): T {
-  return { ...message, ...updates };
-}
-
-function selectedModelForActiveMind(state: AppState, activeMindId: string | null, minds = state.minds): string | null {
-  if (!activeMindId) return null;
-  const selectedModel = minds.find((mind) => mind.mindId === activeMindId)?.selectedModel;
-  if (selectedModel && (state.availableModels.length === 0 || state.availableModels.some((model) => model.id === selectedModel))) {
-    return selectedModel;
-  }
-
-  return state.availableModels[0]?.id ?? null;
-}
-
-function defaultConversationView(): ConversationViewState {
-  return { status: 'idle', streaming: false, modelSwitching: false };
-}
-
-function conversationViewFor(state: AppState, mindId: string): ConversationViewState {
-  return state.conversationViewByMind[mindId] ?? defaultConversationView();
-}
-
-function setConversationView(
-  state: AppState,
-  mindId: string,
-  patch: Partial<ConversationViewState>,
-): Record<string, ConversationViewState> {
-  return {
-    ...state.conversationViewByMind,
-    [mindId]: {
-      ...conversationViewFor(state, mindId),
-      ...patch,
-    },
-  };
-}
-
-function mergeConversationSummaries(
-  existing: ConversationSummary[] | undefined,
-  incoming: ConversationSummary[],
-): ConversationSummary[] {
-  if (!existing?.length) return incoming;
-  const existingById = new Map(existing.map((conversation) => [conversation.sessionId, conversation]));
-  return incoming.map((conversation) => {
-    const current = existingById.get(conversation.sessionId);
-    if (!current) return conversation;
-    const currentUpdatedAt = Date.parse(current.updatedAt);
-    const incomingUpdatedAt = Date.parse(conversation.updatedAt);
-    if (Number.isNaN(currentUpdatedAt) || Number.isNaN(incomingUpdatedAt)) return conversation;
-    return currentUpdatedAt > incomingUpdatedAt ? current : conversation;
-  });
-}
-
-export function handleChatEvent<T extends ChatMessage>(messages: T[], messageId: string, event: ChatEvent): T[] {
-  return messages.map((m) => {
-    if (m.id !== messageId) return m;
-
-    const blocks = [...m.blocks];
-
-    switch (event.type) {
-      case 'chunk': {
-        // Append to last text block, or create one
-        const last = blocks[blocks.length - 1];
-        if (last && last.type === 'text') {
-          blocks[blocks.length - 1] = { ...last, content: last.content + event.content, sdkMessageId: event.sdkMessageId };
-        } else {
-          blocks.push({ type: 'text', sdkMessageId: event.sdkMessageId, content: event.content });
-        }
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'tool_start': {
-        blocks.push({
-          type: 'tool_call',
-          toolCallId: event.toolCallId,
-          toolName: event.toolName,
-          status: 'running',
-          arguments: event.args,
-          parentToolCallId: event.parentToolCallId,
-        });
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'tool_progress': {
-        const idx = blocks.findIndex(b => b.type === 'tool_call' && b.toolCallId === event.toolCallId);
-        if (idx >= 0) {
-          const block = blocks[idx] as Extract<ContentBlock, { type: 'tool_call' }>;
-          blocks[idx] = { ...block, output: (block.output || '') + event.message + '\n' };
-        }
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'tool_output': {
-        const idx = blocks.findIndex(b => b.type === 'tool_call' && b.toolCallId === event.toolCallId);
-        if (idx >= 0) {
-          const block = blocks[idx] as Extract<ContentBlock, { type: 'tool_call' }>;
-          blocks[idx] = { ...block, output: (block.output || '') + event.output };
-        }
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'tool_done': {
-        const idx = blocks.findIndex(b => b.type === 'tool_call' && b.toolCallId === event.toolCallId);
-        if (idx >= 0) {
-          const block = blocks[idx] as Extract<ContentBlock, { type: 'tool_call' }>;
-          blocks[idx] = {
-            ...block,
-            status: event.success ? 'done' : 'error',
-            ...(event.result && { output: (block.output || '') + event.result }),
-            ...(event.error && { error: event.error }),
-          };
-        }
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'permission_request': {
-        // Skip duplicate request events (defensive — SDK shouldn't re-emit
-        // the same requestId, but handle it idempotently anyway).
-        if (blocks.some(b => b.type === 'permission' && b.requestId === event.requestId)) {
-          return m;
-        }
-        blocks.push({
-          type: 'permission',
-          requestId: event.requestId,
-          kind: event.kind,
-          summary: event.summary,
-          outcome: 'pending',
-          ...(event.toolCallId ? { toolCallId: event.toolCallId } : {}),
-        });
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'permission_outcome': {
-        const idx = blocks.findIndex(b => b.type === 'permission' && b.requestId === event.requestId);
-        if (idx >= 0) {
-          const block = blocks[idx] as Extract<ContentBlock, { type: 'permission' }>;
-          blocks[idx] = { ...block, outcome: event.outcome };
-        }
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'reasoning': {
-        const last = blocks[blocks.length - 1];
-        if (last && last.type === 'reasoning' && last.reasoningId === event.reasoningId) {
-          blocks[blocks.length - 1] = { ...last, content: last.content + event.content };
-        } else {
-          blocks.push({ type: 'reasoning', reasoningId: event.reasoningId, content: event.content });
-        }
-        return updateChatMessage(m, { blocks });
-      }
-
-      case 'message_final': {
-        // Reconciliation: add text if this sdkMessageId hasn't been streamed via chunks
-        const hasThisMessage = blocks.some(b => b.type === 'text' && b.sdkMessageId === event.sdkMessageId);
-        if (!hasThisMessage && event.content) {
-          blocks.push({ type: 'text', sdkMessageId: event.sdkMessageId, content: event.content });
-          return updateChatMessage(m, { blocks });
-        }
-        return m;
-      }
-
-      case 'reconnecting':
-        return m; // No-op in blocks — UI uses isStreaming to show indicator
-
-      case 'done':
-        return updateChatMessage(m, { isStreaming: false });
-
-      case 'error':
-        return updateChatMessage(m, {
-          isStreaming: false,
-          blocks: [...blocks, { type: 'text' as const, content: `Error: ${event.message}` }],
-        });
-
-      case 'timeout':
-        return updateChatMessage(m, {
-          isStreaming: false,
-          blocks: [...blocks, { type: 'text' as const, content: `Agent timed out after ${Math.round(event.timeoutMs / 1000)}s` }],
-        });
-
-      default:
-        return m;
-    }
-  });
-}
-
-export function appReducer(state: AppState, action: AppAction): AppState {
-  // Helper: get current mind's messages
-  const activeMsgs = () => state.activeMindId ? (state.messagesByMind[state.activeMindId] ?? []) : [];
-  const setActiveMsgs = (msgs: ChatMessage[]) => {
-    if (!state.activeMindId) return state.messagesByMind;
-    return { ...state.messagesByMind, [state.activeMindId]: msgs };
-  };
-
-  switch (action.type) {
-    case 'ADD_USER_MESSAGE': {
-      const textBlock: ContentBlock = { type: 'text', content: action.payload.content };
-      const blocks: ContentBlock[] = action.payload.images && action.payload.images.length > 0
-        ? [...action.payload.images, textBlock]
-        : [textBlock];
-      // Sending clears only the active mind's draft (#221). Other minds'
-      // unsent drafts remain untouched.
-      const composeDraftByMind = state.activeMindId && state.composeDraftByMind[state.activeMindId]
-        ? (() => {
-            const next = { ...state.composeDraftByMind };
-            delete next[state.activeMindId as string];
-            return next;
-          })()
-        : state.composeDraftByMind;
-      return {
-        ...state,
-        messagesByMind: setActiveMsgs([...activeMsgs(), {
-          id: action.payload.id,
-          role: 'user',
-          blocks,
-          timestamp: action.payload.timestamp,
-        }]),
-        composeDraftByMind,
-      };
-    }
-
-    case 'ADD_ASSISTANT_MESSAGE':
-      if (!state.activeMindId) {
-        return {
-          ...state,
-          isStreaming: true,
-          messagesByMind: setActiveMsgs([...activeMsgs(), {
-            id: action.payload.id,
-            role: 'assistant',
-            blocks: [],
-            timestamp: action.payload.timestamp,
-            isStreaming: true,
-          }]),
-        };
-      }
-      return {
-        ...state,
-        isStreaming: true,
-        streamingByMind: { ...state.streamingByMind, [state.activeMindId]: true },
-        conversationViewByMind: setConversationView(state, state.activeMindId, {
-          status: 'ready',
-          sessionId: state.activeConversationByMind[state.activeMindId] ?? conversationViewFor(state, state.activeMindId).sessionId,
-          pendingSessionId: undefined,
-          streaming: true,
-          error: undefined,
-        }),
-        messagesByMind: setActiveMsgs([...activeMsgs(), {
-          id: action.payload.id,
-          role: 'assistant',
-          blocks: [],
-          timestamp: action.payload.timestamp,
-          isStreaming: true,
-        }]),
-      };
-
-    case 'CHAT_EVENT': {
-      const { mindId, messageId, event } = action.payload;
-      const mindMsgs = state.messagesByMind[mindId] ?? [];
-      const newMessages = handleChatEvent(mindMsgs, messageId, event);
-      const isDone = event.type === 'done' || event.type === 'error' || event.type === 'timeout';
-      const newStreamingByMind = isDone
-        ? { ...state.streamingByMind, [mindId]: false }
-        : state.streamingByMind;
-      return {
-        ...state,
-        messagesByMind: { ...state.messagesByMind, [mindId]: newMessages },
-        isStreaming: isDone ? false : state.isStreaming,
-        streamingByMind: newStreamingByMind,
-        conversationViewByMind: isDone
-          ? setConversationView(state, mindId, {
-            status: 'ready',
-            sessionId: state.activeConversationByMind[mindId] ?? conversationViewFor(state, mindId).sessionId,
-            pendingSessionId: undefined,
-            streaming: false,
-          })
-          : state.conversationViewByMind,
-      };
-    }
-
-    case 'HYDRATE_CHAT_STATE': {
-      const nextConversationViewByMind = action.payload.conversationViewByMind ?? state.conversationViewByMind;
-      const isActiveMindStreaming = state.activeMindId
-        ? Boolean(action.payload.streamingByMind[state.activeMindId] || nextConversationViewByMind[state.activeMindId]?.streaming)
-        : Object.values(action.payload.streamingByMind).some(Boolean);
-      return {
-        ...state,
-        messagesByMind: action.payload.messagesByMind,
-        streamingByMind: action.payload.streamingByMind,
-        conversationViewByMind: nextConversationViewByMind,
-        isStreaming: isActiveMindStreaming,
-      };
-    }
-
-    case 'SET_CONVERSATION_HISTORY': {
-      const conversations = mergeConversationSummaries(
-        state.conversationHistoryByMind[action.payload.mindId],
-        action.payload.conversations,
-      );
-      const activeSessionId = conversations.find((conversation) => conversation.active)?.sessionId;
-      const currentView = conversationViewFor(state, action.payload.mindId);
-      const hasLocalMessages = (state.messagesByMind[action.payload.mindId]?.length ?? 0) > 0;
-      const shouldBindLocalReadyView = currentView.status === 'ready' && currentView.sessionId === undefined && hasLocalMessages;
-      const shouldPreserveView = (currentView.status === 'ready' && currentView.sessionId === activeSessionId)
-        || (currentView.status === 'hydrating' && currentView.pendingSessionId === activeSessionId);
-      return {
-        ...state,
-        conversationHistoryByMind: {
-          ...state.conversationHistoryByMind,
-          [action.payload.mindId]: conversations,
-        },
-        activeConversationByMind: {
-          ...state.activeConversationByMind,
-          [action.payload.mindId]: activeSessionId,
-        },
-        conversationViewByMind: !activeSessionId
-          ? setConversationView(state, action.payload.mindId, {
-            status: 'idle',
-            sessionId: undefined,
-            pendingSessionId: undefined,
-            error: undefined,
-          })
-          : shouldBindLocalReadyView
-            ? setConversationView(state, action.payload.mindId, {
-              status: 'ready',
-              sessionId: activeSessionId,
-              pendingSessionId: undefined,
-              error: undefined,
-            })
-          : !shouldPreserveView
-            ? setConversationView(state, action.payload.mindId, {
-            status: 'idle',
-            sessionId: activeSessionId,
-            pendingSessionId: undefined,
-            error: undefined,
-            })
-            : state.conversationViewByMind,
-      };
-    }
-
-    case 'CONVERSATION_HYDRATING':
-      return {
-        ...state,
-        activeConversationByMind: {
-          ...state.activeConversationByMind,
-          [action.payload.mindId]: action.payload.sessionId,
-        },
-        conversationViewByMind: setConversationView(state, action.payload.mindId, {
-          status: 'hydrating',
-          sessionId: action.payload.sessionId,
-          pendingSessionId: action.payload.sessionId,
-          error: undefined,
-        }),
-      };
-
-    case 'CONVERSATION_HYDRATE_FAILED': {
-      const currentView = conversationViewFor(state, action.payload.mindId);
-      if (currentView.pendingSessionId && currentView.pendingSessionId !== action.payload.sessionId) return state;
-      return {
-        ...state,
-        conversationViewByMind: setConversationView(state, action.payload.mindId, {
-          status: 'idle',
-          sessionId: action.payload.sessionId,
-          pendingSessionId: undefined,
-          error: action.payload.error,
-        }),
-      };
-    }
-
-    case 'RESUME_CONVERSATION': {
-      const currentView = conversationViewFor(state, action.payload.mindId);
-      if (currentView.pendingSessionId && currentView.pendingSessionId !== action.payload.sessionId) return state;
-      return {
-        ...state,
-        messagesByMind: {
-          ...state.messagesByMind,
-          [action.payload.mindId]: action.payload.messages,
-        },
-        conversationHistoryByMind: {
-          ...state.conversationHistoryByMind,
-          [action.payload.mindId]: action.payload.conversations,
-        },
-        activeConversationByMind: {
-          ...state.activeConversationByMind,
-          [action.payload.mindId]: action.payload.sessionId,
-        },
-        streamingByMind: {
-          ...state.streamingByMind,
-          [action.payload.mindId]: false,
-        },
-        conversationViewByMind: setConversationView(state, action.payload.mindId, {
-          status: 'ready',
-          sessionId: action.payload.sessionId,
-          pendingSessionId: undefined,
-          streaming: false,
-          error: undefined,
-        }),
-        isStreaming: state.activeMindId === action.payload.mindId ? false : state.isStreaming,
-      };
-    }
-
-    case 'SET_MINDS':
-      return {
-        ...state,
-        minds: action.payload,
-        selectedModel: selectedModelForActiveMind(state, state.activeMindId, action.payload),
-      };
-
-    case 'SET_AGENT_PROFILE_SUMMARY':
-      return {
-        ...state,
-        agentProfileByMindId: {
-          ...state.agentProfileByMindId,
-          [action.payload.mindId]: action.payload,
-        },
-      };
-
-    case 'SET_ACTIVE_MIND':
-      return {
-        ...state,
-        activeMindId: action.payload,
-        selectedModel: selectedModelForActiveMind(state, action.payload),
-        isStreaming: action.payload ? Boolean(state.streamingByMind[action.payload] || state.conversationViewByMind[action.payload]?.streaming) : false,
-        streamingByMind: state.streamingByMind,
-      };
-
-    case 'ADD_MIND': {
-      const exists = state.minds.some(m => m.mindId === action.payload.mindId);
-      if (exists) return state;
-      return {
-        ...state,
-        minds: [...state.minds, action.payload],
-        activeMindId: state.activeMindId ?? action.payload.mindId,
-      };
-    }
-
-    case 'REMOVE_MIND': {
-      const newMinds = state.minds.filter(m => m.mindId !== action.payload);
-      const newMsgsByMind = { ...state.messagesByMind };
-      const newConversationHistoryByMind = { ...state.conversationHistoryByMind };
-      const newActiveConversationByMind = { ...state.activeConversationByMind };
-      const newConversationViewByMind = { ...state.conversationViewByMind };
-      const newComposeDraftByMind = { ...state.composeDraftByMind };
-      const newAgentProfileByMindId = { ...state.agentProfileByMindId };
-      delete newMsgsByMind[action.payload];
-      delete newConversationHistoryByMind[action.payload];
-      delete newActiveConversationByMind[action.payload];
-      delete newConversationViewByMind[action.payload];
-      delete newComposeDraftByMind[action.payload];
-      delete newAgentProfileByMindId[action.payload];
-      const newActive = state.activeMindId === action.payload
-        ? (newMinds.length > 0 ? newMinds[0].mindId : null)
-        : state.activeMindId;
-      return {
-        ...state,
-        minds: newMinds,
-        activeMindId: newActive,
-        messagesByMind: newMsgsByMind,
-        conversationHistoryByMind: newConversationHistoryByMind,
-        activeConversationByMind: newActiveConversationByMind,
-        conversationViewByMind: newConversationViewByMind,
-        composeDraftByMind: newComposeDraftByMind,
-        agentProfileByMindId: newAgentProfileByMindId,
-        showLanding: newMinds.length === 0,
-      };
-    }
-
-    case 'SET_AVAILABLE_MODELS': {
-      const nextState = { ...state, availableModels: action.payload };
-      return {
-        ...nextState,
-        selectedModel: selectedModelForActiveMind(nextState, nextState.activeMindId),
-      };
-    }
-
-    case 'SET_SELECTED_MODEL':
-      if (
-        state.selectedModel === action.payload
-        && (!state.activeMindId || state.minds.find((mind) => mind.mindId === state.activeMindId)?.selectedModel === (action.payload ?? undefined))
-      ) {
-        return state;
-      }
-      return {
-        ...state,
-        selectedModel: action.payload,
-        minds: state.activeMindId
-          ? state.minds.map((mind) => mind.mindId === state.activeMindId
-            ? { ...mind, selectedModel: action.payload ?? undefined }
-            : mind)
-          : state.minds,
-      };
-
-    case 'SET_MODEL_SWITCHING':
-      return {
-        ...state,
-        conversationViewByMind: setConversationView(state, action.payload.mindId, {
-          modelSwitching: action.payload.switching,
-        }),
-      };
-
-    case 'SET_ACTIVE_VIEW':
-      return { ...state, activeView: action.payload };
-
-    case 'SET_COMPOSE_DRAFT': {
-      const { mindId, draft } = action.payload;
-      const previous = state.composeDraftByMind[mindId] ?? '';
-      if (draft === previous) return state;
-      // Empty drafts must not produce persisted noise (#221) — clear the key
-      // when draft becomes empty so the map stays compact across many minds.
-      if (draft === '') {
-        if (!(mindId in state.composeDraftByMind)) return state;
-        const next = { ...state.composeDraftByMind };
-        delete next[mindId];
-        return { ...state, composeDraftByMind: next };
-      }
-      return {
-        ...state,
-        composeDraftByMind: { ...state.composeDraftByMind, [mindId]: draft },
-      };
-    }
-
-    case 'SET_DISCOVERED_VIEWS':
-      return { ...state, discoveredViews: action.payload };
-
-    case 'SHOW_LANDING':
-      return { ...state, showLanding: true };
-
-    case 'HIDE_LANDING':
-      return { ...state, showLanding: false };
-
-    case 'ACCOUNT_SWITCH_STARTED':
-      return {
-        ...state,
-        runtimePhase: 'switching-account',
-        switchingAccountLogin: action.payload.login,
-        showLanding: false,
-      };
-
-    case 'ACCOUNT_SWITCH_COMPLETED':
-      return {
-        ...state,
-        runtimePhase: 'ready',
-        switchingAccountLogin: null,
-      };
-
-    case 'LOGGED_OUT':
-      return {
-        ...state,
-        runtimePhase: 'ready',
-        switchingAccountLogin: null,
-      };
-
-    case 'MINDS_CHECKED':
-      return { ...state, mindsChecked: true };
-
-    case 'CLEAR_MESSAGES':
-      return {
-        ...state,
-        messagesByMind: state.activeMindId
-          ? { ...state.messagesByMind, [state.activeMindId]: [] }
-          : state.messagesByMind,
-      };
-
-    case 'NEW_CONVERSATION': {
-      const conversationMindId = action.payload?.mindId ?? state.activeMindId;
-      // A new conversation is its own slate — drop any unsent draft for that
-      // mind so the user is not confronted with stale text from the prior
-      // conversation (#221).
-      const composeDraftByMind = conversationMindId && state.composeDraftByMind[conversationMindId]
-        ? (() => {
-            const next = { ...state.composeDraftByMind };
-            delete next[conversationMindId];
-            return next;
-          })()
-        : state.composeDraftByMind;
-      return {
-        ...state,
-        messagesByMind: conversationMindId
-          ? { ...state.messagesByMind, [conversationMindId]: [] }
-          : state.messagesByMind,
-        composeDraftByMind,
-        isStreaming: conversationMindId === state.activeMindId ? false : state.isStreaming,
-        streamingByMind: conversationMindId
-          ? { ...state.streamingByMind, [conversationMindId]: false }
-          : state.streamingByMind,
-        conversationViewByMind: conversationMindId
-          ? setConversationView(state, conversationMindId, {
-            status: 'idle',
-            sessionId: undefined,
-            pendingSessionId: undefined,
-            streaming: false,
-          })
-          : state.conversationViewByMind,
-        chatroomMessages: [],
-        chatroomStreamingByMind: {},
-        chatroomActiveSpeaker: null,
-        chatroomTaskLedger: [],
-        chatroomMetrics: null,
-      };
-    }
-
-    case 'A2A_INCOMING': {
-      const { targetMindId, message, replyMessageId } = action.payload;
-      const targetMsgs = state.messagesByMind[targetMindId] ?? [];
-      const senderMessage: ChatMessage = {
-        id: message.messageId ?? `a2a-${Date.now()}`,
-        role: 'user',
-        blocks: (message.parts ?? []).map((p) => ({
-          type: 'text' as const,
-          content: p.text ?? '',
-        })),
-        timestamp: Date.now(),
-        sender: {
-          mindId: nonEmptyString(message.metadata?.fromId, 'unknown'),
-          name: nonEmptyString(message.metadata?.fromName, 'Unknown Agent'),
-        },
-      };
-      const replyPlaceholder: ChatMessage = {
-        id: replyMessageId,
-        role: 'assistant',
-        blocks: [],
-        timestamp: Date.now(),
-        isStreaming: true,
-      };
-      const isActiveMind = targetMindId === state.activeMindId;
-      return {
-        ...state,
-        messagesByMind: {
-          ...state.messagesByMind,
-          [targetMindId]: [...targetMsgs, senderMessage, replyPlaceholder],
-        },
-        streamingByMind: { ...state.streamingByMind, [targetMindId]: true },
-        isStreaming: isActiveMind ? true : state.isStreaming,
-      };
-    }
-
-    case 'TASK_STATUS_UPDATE': {
-      const TERMINAL_STATES: Set<TaskState> = new Set(['completed', 'failed', 'canceled', 'rejected']);
-      const { taskId, targetMindId, status, contextId } = action.payload;
-      const existingTasks = state.tasksByMind[targetMindId] ?? [];
-      const idx = existingTasks.findIndex(t => t.id === taskId);
-      let updatedTasks: Task[];
-      if (idx >= 0) {
-        const existing = existingTasks[idx];
-        // Don't overwrite terminal tasks with non-terminal status
-        if (TERMINAL_STATES.has(existing.status.state) && !TERMINAL_STATES.has(status.state)) {
-          return state;
-        }
-        updatedTasks = existingTasks.map((t, i) => i === idx ? { ...t, status } : t);
-      } else {
-        const newTask: Task = { id: taskId, contextId, status };
-        updatedTasks = [...existingTasks, newTask];
-      }
-      return {
-        ...state,
-        tasksByMind: { ...state.tasksByMind, [targetMindId]: updatedTasks },
-      };
-    }
-
-    case 'TASK_ARTIFACT_UPDATE': {
-      const { taskId, targetMindId, artifact } = action.payload;
-      const tasks = state.tasksByMind[targetMindId];
-      if (!tasks) return state;
-      const idx = tasks.findIndex(t => t.id === taskId);
-      if (idx < 0) return state;
-      const task = tasks[idx];
-      const updatedTask: Task = { ...task, artifacts: [...(task.artifacts ?? []), artifact] };
-      const updatedTasks = tasks.map((t, i) => i === idx ? updatedTask : t);
-      return {
-        ...state,
-        tasksByMind: { ...state.tasksByMind, [targetMindId]: updatedTasks },
-      };
-    }
-
-    case 'SET_CHATROOM_HISTORY':
-      return { ...state, chatroomMessages: action.payload };
-
-    case 'CHATROOM_USER_MESSAGE':
-      return {
-        ...state,
-        chatroomMessages: [...state.chatroomMessages, action.payload],
-        // Clear stale orchestration state from previous round
-        chatroomActiveSpeaker: null,
-        chatroomMetrics: null,
-        chatroomTaskLedger: [],
-      };
-
-    case 'CHATROOM_AGENT_MESSAGE': {
-      const { messageId, mindId, mindName, roundId, timestamp } = action.payload;
-      const agentMsg: ChatroomMessage = {
-        id: messageId,
-        role: 'assistant',
-        blocks: [],
-        timestamp,
-        isStreaming: true,
-        sender: { mindId, name: mindName },
-        roundId,
-      };
-      return {
-        ...state,
-        chatroomMessages: [...state.chatroomMessages, agentMsg],
-        chatroomStreamingByMind: { ...state.chatroomStreamingByMind, [mindId]: true },
-      };
-    }
-
-    case 'CHATROOM_EVENT': {
-      const { mindId, mindName, messageId, roundId, event } = action.payload;
-
-      // Orchestration events update the active speaker indicator
-      if (isOrchestrationEvent(event)) {
-        switch (event.type) {
-          case 'orchestration:turn-start':
-            return {
-              ...state,
-              chatroomActiveSpeaker: {
-                mindId: event.data.speakerMindId ?? mindId,
-                mindName: event.data.speaker ?? mindName,
-                phase: 'speaking',
-              },
-            };
-
-          case 'orchestration:moderator-decision':
-            return {
-              ...state,
-              chatroomActiveSpeaker: {
-                mindId,
-                mindName,
-                phase: 'moderating',
-              },
-            };
-
-          case 'orchestration:synthesis':
-            return {
-              ...state,
-              chatroomActiveSpeaker: {
-                mindId,
-                mindName,
-                phase: 'synthesizing',
-              },
-            };
-
-          case 'orchestration:convergence':
-          case 'orchestration:handoff-terminated':
-          case 'orchestration:magentic-terminated':
-            return { ...state, chatroomActiveSpeaker: null };
-
-          case 'orchestration:handoff':
-            return {
-              ...state,
-              chatroomActiveSpeaker: {
-                mindId: event.data.toMindId ?? mindId,
-                mindName: event.data.to ?? mindName,
-                phase: 'speaking',
-              },
-            };
-
-          case 'orchestration:manager-plan':
-          case 'orchestration:task-ledger-update':
-            return {
-              ...state,
-              chatroomActiveSpeaker: {
-                mindId,
-                mindName,
-                phase: 'moderating',
-              },
-              ...(event.type === 'orchestration:task-ledger-update' && event.data.ledger
-                ? { chatroomTaskLedger: event.data.ledger as TaskLedgerItem[] }
-                : {}),
-            };
-
-          case 'orchestration:metrics':
-            return {
-              ...state,
-              chatroomMetrics: event.data as AppState['chatroomMetrics'],
-            };
-
-          case 'orchestration:approval-requested':
-          case 'orchestration:approval-decided':
-          default:
-            return state;
-        }
-      }
-
-      // At this point, event is a ChatEvent (not an OrchestrationEvent)
-      const chatEvent = event as ChatEvent;
-
-      let messages = state.chatroomMessages;
-
-      // Auto-create placeholder if this is the first event for an unknown message
-      const exists = messages.some(m => m.id === messageId);
-      if (!exists) {
-        const placeholder: ChatroomMessage = {
-          id: messageId,
-          role: 'assistant',
-          blocks: [],
-          timestamp: Date.now(),
-          isStreaming: true,
-          sender: { mindId, name: mindName },
-          roundId,
-        };
-        messages = [...messages, placeholder];
-      }
-
-      const newMessages = handleChatEvent(messages, messageId, chatEvent);
-      const isDone = chatEvent.type === 'done' || chatEvent.type === 'error' || chatEvent.type === 'timeout';
-      return {
-        ...state,
-        chatroomMessages: newMessages,
-        chatroomStreamingByMind: isDone
-          ? { ...state.chatroomStreamingByMind, [mindId]: false }
-          : { ...state.chatroomStreamingByMind, [mindId]: true },
-        // Clear active speaker when the active speaker finishes
-        ...(isDone && state.chatroomActiveSpeaker?.mindId === mindId
-          ? { chatroomActiveSpeaker: null }
-          : {}),
-      };
-    }
-
-    case 'CHATROOM_CLEAR':
-      return { ...state, chatroomMessages: [], chatroomStreamingByMind: {}, chatroomActiveSpeaker: null, chatroomTaskLedger: [], chatroomMetrics: null };
-
-    case 'SET_CHATROOM_TASK_LEDGER':
-      return { ...state, chatroomTaskLedger: action.payload };
-
-    case 'SET_ORCHESTRATION':
-      return { ...state, chatroomOrchestration: action.payload };
-
-    case 'SET_GROUP_CHAT_CONFIG':
-      return { ...state, chatroomGroupChatConfig: action.payload };
-
-    case 'SET_HANDOFF_CONFIG':
-      return { ...state, chatroomHandoffConfig: action.payload };
-
-    case 'SET_MAGENTIC_CONFIG':
-      return { ...state, chatroomMagenticConfig: action.payload };
-
-    case 'CHATROOM_ACTIVE_SPEAKER':
-      return { ...state, chatroomActiveSpeaker: action.payload };
-
-    case 'SET_CHATROOM_DISABLED_MIND_IDS':
-      return { ...state, chatroomDisabledMindIds: action.payload };
-
-    default:
-      return state;
-  }
-}
+export { appReducer, handleChatEvent, getPlainContent } from './reducers';
